+++ app/Http/Controllers/LA/CountriesController.php
@@ -81,7 +81,7 @@ class CountriesController extends Controller
                         }
 
                         if(Module::hasAccess("Countries", "delete")) {
-                            $output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.countries.destroy', $values[$i]->id], 'method' => 'delete', 'style'=>'display:inline']);
+                            $output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.countries.destroy', $values[$i]->id], 'method' => 'delete', 'style'=>'display:inline', 'class'=>'custom-confirm']);
                             $output .= ' <button class="btn btn-danger btn-xs" type="submit"><i class="fa fa-times"></i></button>';
                             $output .= Form::close();
                         }
