+++ app/Http/Controllers/LA/Jobs_ADsController.php
@@ -88,7 +88,7 @@ class Jobs_ADsController extends Controller
                         }
 
                         if(Module::hasAccess("Jobs_ADs", "delete")) {
-                            $output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.jobs_ads.destroy', $values[$i]->id], 'method' => 'delete', 'style'=>'display:inline']);
+                            $output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.jobs_ads.destroy', $values[$i]->id], 'method' => 'delete', 'style'=>'display:inline', 'class'=>'custom-confirm']);
                             $output .= ' <button class="btn btn-danger btn-xs" type="submit"><i class="fa fa-times"></i></button>';
                             $output .= Form::close();
                         }
