+++ app/Http/Controllers/LA/NotificationsController.php
@@ -91,7 +91,7 @@ class NotificationsController extends Controller
                         }
 
                         if(Module::hasAccess("Notifications", "delete")) {
-                            $output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.notifications.destroy', $values[$i]->id], 'method' => 'delete', 'style'=>'display:inline']);
+                            $output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.notifications.destroy', $values[$i]->id], 'method' => 'delete', 'style'=>'display:inline', 'class'=>'custom-confirm']);
                             $output .= ' <button class="btn btn-danger btn-xs" type="submit"><i class="fa fa-times"></i></button>';
                             $output .= Form::close();
                         }
