+++ app/Http/Controllers/LA/PermissionsController.php
@@ -241,7 +241,7 @@ class PermissionsController extends Controller
 				}
 				
 				if(Module::hasAccess("Permissions", "delete")) {
-					$output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.permissions.destroy', $data->data[$i][0]], 'method' => 'delete', 'style'=>'display:inline']);
+					$output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.permissions.destroy', $data->data[$i][0]], 'method' => 'delete', 'style'=>'display:inline', 'class'=>'custom-confirm']);
 					$output .= ' <button class="btn btn-danger btn-xs" type="submit"><i class="fa fa-times"></i></button>';
 					$output .= Form::close();
 				}
