+++ app/Http/Controllers/LA/User_certificationsController.php
@@ -111,7 +111,7 @@ class User_certificationsController extends Controller
                         }
 
                         if (Module::hasAccess("User_certifications", "delete")) {
-                            $output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.user_certifications.destroy', $values[$i]->id], 'method' => 'delete', 'style' => 'display:inline']);
+                            $output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.user_certifications.destroy', $values[$i]->id], 'method' => 'delete', 'style' => 'display:inline', 'class'=>'custom-confirm']);
                             $output .= ' <button class="btn btn-danger btn-xs" type="submit"><i class="fa fa-times"></i></button>';
                             $output .= Form::close();
                         }
