+++ app/Http/Controllers/LA/Validation_statusesController.php
@@ -236,7 +236,7 @@ class Validation_statusesController extends Controller
 				}
 				
 				if(Module::hasAccess("Validation_statuses", "delete")) {
-					$output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.validation_statuses.destroy', $data->data[$i][0]], 'method' => 'delete', 'style'=>'display:inline']);
+					$output .= Form::open(['route' => [config('laraadmin.adminRoute') . '.validation_statuses.destroy', $data->data[$i][0]], 'method' => 'delete', 'style'=>'display:inline', 'class'=>'custom-confirm']);
 					$output .= ' <button class="btn btn-danger btn-xs" type="submit"><i class="fa fa-times"></i></button>';
 					$output .= Form::close();
 				}
